[feat] change classes to models file
