--- player.py
@@ -1,4 +1,0 @@
-class Player:
-    def __init__(self, nome):
-        self.nome: str = nome
-        self.pokemon: Pokemon = []
